+++ packages/insomnia/src/ui/routes/project.tsx
@@ -22,7 +22,6 @@ import {
   useLoaderData,
   useNavigate,
   useParams,
-  useRouteLoaderData,
   useSearchParams,
 } from 'react-router-dom';
 
@@ -65,7 +64,7 @@ import { ImportModal } from '../components/modals/import-modal';
 import { EmptyStatePane } from '../components/panes/project-empty-state-pane';
 import { SidebarLayout } from '../components/sidebar-layout';
 import { TimeFromNow } from '../components/time-from-now';
-import { RootLoaderData } from './root';
+import { useOrganizationLoaderData } from './organization';
 
 export interface WorkspaceWithMetadata {
   _id: string;
@@ -323,7 +322,7 @@ const ProjectRoute: FC = () => {
     projectId: string;
   };
 
-  const { organizations } = useRouteLoaderData('root') as RootLoaderData;
+  const { organizations } = useOrganizationLoaderData();
 
   const [searchParams, setSearchParams] = useSearchParams();
   const [isGitRepositoryCloneModalOpen, setIsGitRepositoryCloneModalOpen] =
